Move ToolbarExtension out of plugins/ directory
